made command to move vomitElevator back down
