+++ src/org/usfirst/frc4904/robot/commands/DumpUp.java
@@ -1,12 +1,12 @@
-package org.usfirst.frc4904.robot.commmands;
+package org.usfirst.frc4904.robot.commands;
 
 
 import org.usfirst.frc4904.robot.RobotMap;
 import edu.wpi.first.wpilibj.command.Command;
 
-public class DumpCommandUp extends Command {
+public class DumpUp extends Command {
 	
-	public DumpCommandUp() {
+	public DumpUp() {
 		requires(RobotMap.Component.ballDumper);
 	}
